refactor(runner): extract linkPinnedPackages helper from innerHop
